Create toggle command, separate actions into functions.
Add toggle command: AppPauser already keeps track of the application
status, so making it possible to toggle that is a must.

Separate actions into functions: Separate functions for Pause and
Resume make the code more modular. Error verification was also added
to these functions, in order to detect if some condition made it not
possible to send signals to the controlled application.